@@ -1,6 +1,6 @@
 {
   "name": "bwcallsettingsweb",
-  "version": "0.0.1",
+  "version": "0.0.0",
   "license": "MIT",
   "angular-cli": {},
   "scripts": {
@@ -8,42 +8,52 @@
     "start": "ng serve",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "build": "ng build"
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "4.3.1",
-    "@angular/compiler": "4.3.1",
-    "@angular/core": "4.3.1",
-    "@angular/forms": "4.3.1",
-    "@angular/http": "4.3.1",
-    "@angular/material": "2.0.0-beta.2",
-    "@angular/platform-browser": "4.3.1",
-    "@angular/platform-browser-dynamic": "4.3.1",
-    "@angular/router": "4.3.1",
-    "core-js": "2.4.1",
+    "@angular/animations": "^7.2.3",
+    "@angular/cdk": "^7.3.0",
+    "@angular/common": "^7.2.3",
+    "@angular/compiler": "^7.2.3",
+    "@angular/core": "^7.2.3",
+    "@angular/forms": "^7.2.3",
+    "@angular/http": "^7.2.3",
+    "@angular/material": "^7.3.2",
+    "@angular/material-moment-adapter": "^7.3.3",
+    "@angular/platform-browser": "^7.2.3",
+    "@angular/platform-browser-dynamic": "^7.2.3",
+    "@angular/platform-server": "^7.2.3",
+    "@angular/router": "^7.2.3",
+    "core-js": "2.6.3",
     "hammerjs": "2.0.8",
-    "rxjs": "5.4.2",
-    "zone.js": "0.8.5"
+    "moment": "^2.24.0",
+    "ngx-webstorage": "^3.0.2",
+    "rxjs": "6.4.0",
+    "rxjs-compat": "^6.4.0",
+    "tslib": "^1.9.0",
+    "zone.js": "0.8.29"
   },
   "devDependencies": {
-    "@angular/cli": "^1.2.4",
-    "@angular/compiler-cli": "4.3.1",
-    "@types/hammerjs": "2.0.34",
-    "@types/jasmine": "2.5.38",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~2.0.0-beta.4",
-    "jasmine-core": "~2.5.2",
-    "jasmine-spec-reporter": "~3.2.0",
-    "karma": "~1.4.1",
-    "karma-chrome-launcher": "~2.0.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "0.2.0",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "0.2.2",
-    "protractor": "~5.1.0",
-    "ts-node": "~2.0.0",
-    "tslint": "~4.4.2",
-    "typescript": "^2.4.2"
+    "@angular-devkit/build-angular": "~0.13.0",
+    "@angular/cli": "^7.3.0",
+    "@angular/compiler-cli": "^7.2.3",
+    "@types/hammerjs": "2.0.36",
+    "@types/jasmine": "3.3.8",
+    "@types/node": "~10.12.21",
+    "codelyzer": "~4.5.0",
+    "jasmine-core": "~3.3.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.0.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-cli": "~2.0.0",
+    "karma-coverage-istanbul-reporter": "2.0.4",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "1.4.0",
+    "protractor": "~5.4.2",
+    "ts-node": "~8.0.2",
+    "tslint": "~5.12.1",
+    "typescript": "3.1.6"
   }
 }
