@@ -2,7 +2,7 @@
 
 import { Injectable, Inject, Component } from '@angular/core';
 import { Http, Response, Request, Headers, RequestOptions } from '@angular/http';
-import { Observable } from "rxjs/Observable";
+import { Observable } from 'rxjs/Observable';
 import { HttpServices } from 'app/AppCommon/httpservices.service';
 import { CallCenterServiceInput } from 'app/CallControl/CallCenterQueues/callcenterServiceInput.service';
 import { CallCenterArray } from 'app/CallControl/CallCenterQueues/callcenterServiceInput.service';
@@ -20,34 +20,32 @@ export class CallCenterService {
     getCallCenterService(ccUrl, postCCGet) {
         this.httpservices.httpGetRequest(ccUrl)
             .subscribe((res) => {
-                var ccParsedJson = res.json();
-                if(ccParsedJson["CallCenter"]["agentACDState"]){
-                var acdStateInput = ccParsedJson["CallCenter"]["agentACDState"]["$"];
+                let ccParsedJson = res.json();
+                if (ccParsedJson['CallCenter']['agentACDState']) {
+                    let acdStateInput = ccParsedJson['CallCenter']['agentACDState']['$'];
 
-                if (acdStateInput == "Sign-In") {
-                    acdStateInput = this.customizedTextJson.callcenter.sign_in;
-                }
-                else if (acdStateInput == "Sign-Out") {
-                    acdStateInput = this.customizedTextJson.callcenter.sign_out;
-                }
-                else if (acdStateInput == "Wrap-Up") {
-                    acdStateInput = this.customizedTextJson.callcenter.wrap_up;
-                } else if (acdStateInput == "Available") {
-                    acdStateInput = this.customizedTextJson.callcenter.available;
-                } else if (acdStateInput == "Unavailable") {
-                    acdStateInput = this.customizedTextJson.callcenter.unavailable;
-                }
+                    if (acdStateInput === 'Sign-In') {
+                        acdStateInput = this.customizedTextJson.callcenter.sign_in;
+                    } else if (acdStateInput === 'Sign-Out') {
+                        acdStateInput = this.customizedTextJson.callcenter.sign_out;
+                    } else if (acdStateInput === 'Wrap-Up') {
+                        acdStateInput = this.customizedTextJson.callcenter.wrap_up;
+                    } else if (acdStateInput === 'Available') {
+                        acdStateInput = this.customizedTextJson.callcenter.available;
+                    } else if (acdStateInput === 'Unavailable') {
+                        acdStateInput = this.customizedTextJson.callcenter.unavailable;
+                    }
 
-                this.callcenterServiceInput.setAcdState(acdStateInput);
+                    this.callcenterServiceInput.setAcdState(acdStateInput);
                 } else {
-                    this.callcenterServiceInput.setAcdState("");
+                    this.callcenterServiceInput.setAcdState('');
                 }
                 this.callcenterServiceInput.setCallcenter(ccParsedJson);
-                  if (ccParsedJson.CallCenter.agentUnavailableCode) {
-                this.callcenterServiceInput.setSelectedUnavailableCode(ccParsedJson.CallCenter.agentUnavailableCode.$);
-                  } else {
-                      this.callcenterServiceInput.setSelectedUnavailableCode(null);
-                  }
+                if (ccParsedJson.CallCenter.agentUnavailableCode) {
+                    this.callcenterServiceInput.setSelectedUnavailableCode(ccParsedJson.CallCenter.agentUnavailableCode.$);
+                } else {
+                    this.callcenterServiceInput.setSelectedUnavailableCode(null);
+                }
                 postCCGet(null);
             }, (err) => {
                 postCCGet(err);
@@ -58,8 +56,8 @@ export class CallCenterService {
 
         this.httpservices.httpGetRequest(unavailableCodeUrl)
             .subscribe((res) => {
-                var ucParsedJson = (res.json());
-                this.callcenterServiceInput.setIsUCenabled(ucParsedJson.ACDAgentUnavailableCodes.enableAgentUnavailableCodes.$ == "true");
+                let ucParsedJson = (res.json());
+                this.callcenterServiceInput.setIsUCenabled(ucParsedJson.ACDAgentUnavailableCodes.enableAgentUnavailableCodes.$ === 'true');
 
                 if (this.callcenterServiceInput.getIsUCenabled()) {
                     this.callcenterServiceInput.setUnavailableCodeArray(ucParsedJson);
@@ -75,7 +73,7 @@ export class CallCenterService {
     getCallCenterName(ccNameUrl, index, postCCnameGet) {
         this.httpservices.httpGetRequest(ccNameUrl)
             .subscribe((res) => {
-                var ccNameParsedJson = (res.json());
+                let ccNameParsedJson = (res.json());
                 this.callcenterServiceInput.setCallCenterName(ccNameParsedJson.ACDProfile.serviceInstanceProfile.name.$, index);
 
                 postCCnameGet(null);
@@ -86,23 +84,23 @@ export class CallCenterService {
     }
 
     putAcdState(callCenterUrl, acdState, isUnavailableStateSet, isUCenabled, unavailableCode, postAcdStatePut) {
-        var body = '<?xml version="1.0" encoding="UTF-8"?><CallCenter xmlns="http://schema.broadsoft.com/xsi">';
+        let body = '<?xml version="1.0" encoding="UTF-8"?><CallCenter xmlns="http://schema.broadsoft.com/xsi">';
 
-        var acd;
+        let acd;
 
-        if (acdState == this.customizedTextJson.callcenter.sign_in) {
-            acd = "Sign-In";
+        if (acdState === this.customizedTextJson.callcenter.sign_in) {
+            acd = 'Sign-In';
         }
-        else if (acdState == this.customizedTextJson.callcenter.sign_out) {
-            acd = "Sign-Out";
+        else if (acdState === this.customizedTextJson.callcenter.sign_out) {
+            acd = 'Sign-Out';
         }
-        else if (acdState == this.customizedTextJson.callcenter.wrap_up) {
-            acd = "Wrap-Up";
-        } else if (acdState == this.customizedTextJson.callcenter.available) {
-            acd = "Available";
-        } else if (acdState == this.customizedTextJson.callcenter.unavailable) {
-            acd = "Unavailable";
-        } 
+        else if (acdState === this.customizedTextJson.callcenter.wrap_up) {
+            acd = 'Wrap-Up';
+        } else if (acdState === this.customizedTextJson.callcenter.available) {
+            acd = 'Available';
+        } else if (acdState === this.customizedTextJson.callcenter.unavailable) {
+            acd = 'Unavailable';
+        }
 
         body = body + '<agentACDState>' + acd + '</agentACDState>';
 
@@ -126,7 +124,7 @@ export class CallCenterService {
 
     putCCData(callCenterUrl, ccElement, postCCput) {
 
-        var body = '<?xml version="1.0" encoding="UTF-8"?><CallCenter xmlns="http://schema.broadsoft.com/xsi"><callCenterList><callCenterDetails><serviceUserId>';
+        let body = '<?xml version="1.0" encoding="UTF-8"?><CallCenter xmlns="http://schema.broadsoft.com/xsi"><callCenterList><callCenterDetails><serviceUserId>';
 
         body = body + ccElement.callCenterId + '</serviceUserId><available>' + ccElement.active + '</available></callCenterDetails></callCenterList></CallCenter>';
 
@@ -134,7 +132,7 @@ export class CallCenterService {
             .subscribe((res) => {
                 postCCput(null);
             }, (err) => {
-                            ccElement.setCallCenterActive(!ccElement.getCallCenterActive());
+                ccElement.setCallCenterActive(!ccElement.getCallCenterActive());
                 postCCput(err);
             });
 
