@@ -1,4 +1,4 @@
-/* Copyright � 2017 BroadSoft Inc. */
+/* Copyright © 2017 BroadSoft Inc. */
 package com.broadsoft.callsettingsweb.server.util;
 
 import java.io.BufferedReader;
